requirements.txt
discord.py==1.7.3
requests==2.26.0
python-dotenv==0.19.0